test: refactor cluster and replicaset
Separate different command tests, use introduced helpers.
