+++ test/conftest.py
@@ -5,6 +5,7 @@ import tempfile
 
 import py
 import pytest
+from etcd_helper import EtcdInstance
 
 from utils import create_tt_config
 
@@ -80,3 +81,21 @@ def tmpdir_with_tarantool(tt_cmd, request):
     assert tt_process.returncode == 0
 
     return tmpdir
+
+
+@pytest.fixture(scope="session")
+def etcd_session(request, session_tmpdir):
+    tmpdir = session_tmpdir
+    host = "localhost"
+    port = 12388
+    etcd_instance = EtcdInstance(host, port, tmpdir)
+    etcd_instance.start()
+
+    request.addfinalizer(lambda: etcd_instance.stop())
+    return etcd_instance
+
+
+@pytest.fixture(scope="function")
+def etcd(etcd_session):
+    etcd_session.truncate()
+    return etcd_session
